// clients/action-types.ts
export enum ClientsActionTypes {
  SET_CLIENTS_LIST = "SET_CLIENTS_LIST",
  SET_PREPARE_DATA_STATUS = "SET_PREPARE_DATA_STATUS",
  SET_HAS_USER_ESD = "SET_HAS_USER_ESD",
  SET_ESD_LOADING_STATUS = "SET_ESD_LOADING_STATUS",
  SET_AVAILABLE_OPERATIONS = "SET_AVAILABLE_OPERATIONS",
  SET_TYPE_USER_SIGNING = "SET_TYPE_USER_SIGNING",
}